docs(readme): adapt to use case
Since this project stopped before it reached a state where it serves as
more than a playground and reference to what I learned in d3 it will
stay this way now for future reference on d3, scatterplot matrices and
parallel coordinates. Therefore this update to the project README.

Adding also proper attribution to supporters and related projects.